requirements.txt
Django ~= 3.1.0
beautifulsoup4
django-compressor
django-haystack
django-user-agents
django-widget-tweaks
elasticsearch ~= 5.0
lxml
Markdown
Pillow
psycopg2
requests
slimit
